Moved the tests from TestOzoneShellEC to TestOzoneShellHA to reuse the minicluster
